Prevent username colisions
Discord merges together webhook messages with the same username, so two users that happen to have the same name on WhatsApp would show up as the same person on Discord, causing confusion. This commit adds an option that adds a bunch of invisible characters to the name based on the author id, to hopefully prevent conflicts. Since Discord doesn't allow trailing whitespace, a small character is appended at the very end of the username.